Draft minimum viable product of Chivolo app
Created index.html, index.css, and index.js to draft the MVP of kid's video application.
This one-page app contains a logo at the top, followed by an iframe  video player that plays the
first video in the Chivolo Main Page playlist, followed by image-title pairs that the
user can click to load the associated video in the video player.